Moved util folder and misc cleanup
