+++ package.json
@@ -26,7 +26,8 @@
     "@reduxjs/toolkit": "^1.9.5",
     "@shopify/flash-list": "1.4.3",
     "@supabase/supabase-js": "^2.38.4",
-    "@tamagui/config": "^1.81.4",
+    "@tamagui/config": "^1.84.0",
+    "@tamagui/lucide-icons": "^1.84.0",
     "burnt": "^0.12.1",
     "expo": "~49.0.3",
     "expo-application": "~5.3.0",
@@ -56,12 +57,12 @@
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-native-svg": "13.9.0",
+    "react-native-svg": "^13.9.0",
     "react-native-url-polyfill": "^2.0.0",
     "react-native-web": "~0.19.6",
     "react-redux": "^8.1.1",
     "reanimated-color-picker": "^2.4.1",
-    "tamagui": "^1.81.4",
+    "tamagui": "^1.84.0",
     "zod": "^3.21.4"
   },
   "devDependencies": {
