+++ src/components/molecules/TagRow.tsx
@@ -1,4 +1,4 @@
-import { hexToRGBA } from "@/lib/util/color";
+import { hexToRGBA } from "@/util/color";
 import { HStack, IPressableProps, Pressable, Square, Text } from "native-base";
 
 interface TagListRowProps extends IPressableProps {
