--- src/lib/util/index.ts
@@ -1,3 +1,0 @@
-export function generateShortUUID() {
-	return Math.random().toString(36).substring(2, 15);
-}
