+++ src/store/user/user.interface.ts
@@ -3,10 +3,9 @@ import { z } from "zod";
 const appSubscription = z.object({
 	subscribed: z.boolean(),
 });
-
 export type AppSubscription = z.infer<typeof appSubscription>;
 
-export const SpotifyToken = z.object({
+export const spotifyToken = z.object({
 	accessToken: z.string(),
 	tokenType: z.string(),
 	expiresIn: z.number().optional(),
@@ -14,26 +13,22 @@ export const SpotifyToken = z.object({
 	idToken: z.string().optional(),
 	issuedAt: z.number().optional(),
 });
+export type SpotifyToken = z.infer<typeof spotifyToken>;
 
-export type SpotifyToken = z.infer<typeof SpotifyToken>;
-
-const User = z.object({
+const user = z.object({
 	id: z.string(),
 });
-
-export type User = z.infer<typeof User>;
+export type User = z.infer<typeof user>;
 
 const spotifyUser = z.object({
 	id: z.string(),
 });
-
 export type SpotifyUser = z.infer<typeof spotifyUser>;
 
-const UserState = z.object({
-	user: User.optional(),
-	spotifyToken: SpotifyToken.optional(),
+const userState = z.object({
+	user: user.optional(),
+	spotifyToken: spotifyToken.optional(),
 	appSubscription: appSubscription.optional(),
 	spotifyUser: spotifyUser.optional(),
 });
-
-export type UserState = z.infer<typeof UserState>;
+export type UserState = z.infer<typeof userState>;
