+++ src/types/env.ts
@@ -2,7 +2,6 @@ import { z } from "zod";
 
 const envVariables = z.object({
 	EXPO_PUBLIC_APTABASE_KEY: z.string(),
-	EXPO_PUBLIC_SENTRY_DSN: z.string(),
 	EXPO_PUBLIC_SUPABASE_URL: z.string(),
 	EXPO_PUBLIC_SUPABASE_KEY: z.string(),
 	EXPO_PUBLIC_SPOTIFY_CLIENT_ID: z.string(),
